Refactor: CombatRestrictions 패키지 옮김
