+++ src/main/java/com/dace/dmgr/combat/entity/module/MoveModule.java
@@ -2,7 +2,7 @@ package com.dace.dmgr.combat.entity.module;
 
 import com.dace.dmgr.combat.entity.Damageable;
 import com.dace.dmgr.combat.entity.Movable;
-import com.dace.dmgr.combat.entity.module.statuseffect.CombatRestrictions;
+import com.dace.dmgr.combat.entity.CombatRestrictions;
 import com.dace.dmgr.user.User;
 import com.dace.dmgr.util.task.IntervalTask;
 import com.dace.dmgr.util.task.TaskUtil;
