+++ src/main/java/com/dace/dmgr/combat/entity/module/statuseffect/Invulnerable.java
@@ -1,6 +1,7 @@
 package com.dace.dmgr.combat.entity.module.statuseffect;
 
 import com.dace.dmgr.combat.entity.CombatEntity;
+import com.dace.dmgr.combat.entity.CombatRestrictions;
 import com.dace.dmgr.combat.entity.Damageable;
 import lombok.AccessLevel;
 import lombok.Getter;
